Limit files added to /boot
New versions add files for the Raspberry Pi 4 that when added exceed the 30MB
boot partition for RancherOS. The plain set boots standard config, and the
"cd" versions are cutdown versions for gpu_mem=16 that disables some GPU
features. Removes the debug (db) and extended (x) versions, debug versions
are less common, but the x versisons are required for camera use.